requirements.txt
web3
eth_account
python-dotenv
solcx
